task 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 int main()
 {
-	unsigned long long int a = 1;
-	unsigned long long int b = 2;
+	long int a = 1;
+	long int b = 2;
 	int n = 50;
-	unsigned long long int tmp;
+	long int tmp;
 
 	while (n > 0)
 	{
-		printf("%llu", a);
+		printf("%lu", a);
 		tmp = a;
 		if (n > 1)
 		{
